backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]
payments-py[fastapi]
mcp>=1.26.0
exa-py>=2.7.0
anthropic>=0.40.0
supabase>=2.28.0
pydantic>=2.0.0
python-dotenv
httpx>=0.27.0
apify-client>=1.8.0
